+++ spiridonov.mikhail/P4/minSumMgt.cpp
@@ -1,23 +1,23 @@
 #include "getMinimumSum.hpp"
 #include "minSumMgt.hpp"
 
-void spiridonov::readMatrix(std::ifstream& input, int* matrix, int size)
+void spiridonov::readMatrix(std::ifstream& input, int* matrix, size_t size)
 {
-  for (int i = 0; i < size; i++)
+  for (size_t i = 0; i < size; i++)
   {
     input >> matrix[i];
   }
 }
 
-int spiridonov::getMinimumSum(int * matrix, int rows, int cols)
+size_t spiridonov::getMinimumSum(int * matrix, long long rows, long long cols)
 {
   int min_sum = matrix[cols - 1];
 
-  for (int i = 1; i < rows; i++)
+  for (long long i = 1; i < rows; i++)
   {
     int sum = matrix[i * cols + cols - 1];
 
-    for (int j = 0; j < cols && i + j < rows; j++)
+    for (long long j = 0; j < cols && i + j < rows; j++)
     {
       sum += matrix[(i + j) * cols + cols - 1 - j];
     }
@@ -28,11 +28,11 @@ int spiridonov::getMinimumSum(int * matrix, int rows, int cols)
     }
   }
 
-  for (int i = 1; i < cols - 1; i++)
+  for (long long i = 1; i < cols - 1; i++)
   {
     int sum = matrix[i + (cols - 1) * cols];
 
-    for (int j = 0; j < rows && i + j < cols; j++)
+    for (long long j = 0; j < rows && i + j < cols; j++)
     {
       sum += matrix[(rows - 1 - j) * cols + i + j];
     }
